+++ tests/consumers/CMakeLists.txt
@@ -10,7 +10,7 @@ set(SOURCES
   ${dir}/spectrum_time.cpp
   ${dir}/stats_scalar.cpp
   ${dir}/time_delta_1d.cpp
-  ${dir}/time_domain.cpp
+  ${dir}/TimeDomainTest.cpp
   ${dir}/TOF1DTest.cpp
   ${dir}/TOF1DCorrelateTest.cpp
   ${dir}/TOFVal2DTest.cpp
